+++ docs/source/changelog.rst
@@ -31,6 +31,12 @@ Glossary
 ----------------------
 Releases
 ----------------------
+v2.4
+=============
+- Multiple accounts support
+- |BREAK_CH| :func:`~daf.core.add_object` and :func:`~daf.core.remove_object` functions no longer accepts API wrapper objects or ``int`` type for the ``snowflake`` parameter.
+- |BREAK_CH| Removed ``daf.core.get_guild_user`` function due to multiple accounts support.
+- |BREAK_CH| Removed ``daf.client.get_client`` function due to multiple accounts support.
 
 v2.3
 =============